feat: UX 향상, 삭제 시 팝업 띄워서 확인

+++ lib/home_page.dart
@@ -1,4 +1,5 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_todo_app/todo.dart';
 import 'package:flutter_todo_app/todo_widget.dart';
@@ -77,7 +78,7 @@ class _HomePageState extends State<HomePage> {
         itemBuilder: (context, index) {
           Todo todoItem = todoList[index];
           return GestureDetector(
-            onTap: () async{
+            onTap: () async {
               print('투두 위젯 터치됨');
 
               // 1. 파이어베이스 콘솔에서 파이어 스토어 클릭하기 => Firestore 인스턴스 가져오기
@@ -87,24 +88,59 @@ class _HomePageState extends State<HomePage> {
               // 3. 수정할 문서 클릭 => 문서 참조 만들기
               final docRef = colRef.doc(todoItem.Id);
               // 4. 수정 => 문서 참조 객체 이용해서 update 함수 호출
-              await docRef.update({
-                'isDone' : !todoItem.isDone
-              });
+              await docRef.update({'isDone': !todoItem.isDone});
               // 데이터 새로고침
               loadTodoList();
             },
-            onLongPress: () async{
+            onLongPress: () async {
               print('길게 터치됨');
-              // 1. 파이어스토어 인스턴스 가져오기
-               final firestore = FirebaseFirestore.instance;
-              // 2. 컬렉션 참조 만들기
+              final result = await showCupertinoDialog(
+                context: context,
+                builder: (context) {
+                  return CupertinoAlertDialog(
+                    title: Text('삭제하시겠습니까?'),
+                    actions: [
+                      CupertinoDialogAction(
+                          onPressed: () {
+                            Navigator.pop(context, true);
+                          },
+                          child: Text(
+                            '삭제',
+                            style: TextStyle(color: Colors.red),
+                          )),
+                      CupertinoDialogAction(
+                          onPressed: () {
+                            Navigator.pop(context, false);
+                          },
+                          child: Text(
+                            '취소',
+                            style: TextStyle(color: Colors.blue),
+                          )),
+                    ],
+                  );
+                  // 정렬해주는 위젯 없으면 꽉참
+                  return Center(
+                    child: Container(
+                      width: 300,
+                      height: 300,
+                      color: Colors.amber,
+                    ),
+                  );
+                },
+              );
+              print('팝업 닫힘');
+              if (result == true) {
+                // 1. 파이어스토어 인스턴스 가져오기
+                final firestore = FirebaseFirestore.instance;
+                // 2. 컬렉션 참조 만들기
                 final colRef = firestore.collection('todos');
-              // 3. 문서 참조 만들기
-              final docRef = colRef.doc(todoItem.Id);
-              // 4. 삭제
-              await docRef.delete();
-              // 데이터 새로고침
-              loadTodoList();
+                // 3. 문서 참조 만들기
+                final docRef = colRef.doc(todoItem.Id);
+                // 4. 삭제
+                await docRef.delete();
+                // 데이터 새로고침
+                loadTodoList();
+              }
             },
             child: TodoWidget(
               title: todoItem.title,
